requirements.txt
# Core Dependencies
streamlit==1.55.0
langchain==1.2.12
langchain-community==0.4.1
langchain-core==1.2.19
langchain-google-genai==4.2.1
langchain-huggingface==1.2.1
langchain-chroma==1.1.0
chromadb==1.5.5
pdfplumber==0.11.9
python-dotenv==1.2.2
sentence-transformers==5.3.0
pandas==2.3.3
langfuse==4.0.0